[FEAT] Add Task Manager Feature
As per previously discussed, we intended to implement a Task manager
feature. This is the commit(s) that will implement the feature.

Included in this PR/commit(s) are tests as well as new classes that
are related to the Task Manager feature.

Most of it is a replication of how ContactList interacts with the
various Logic/Storage/Model Managers.

Note that UI has NOT been touched yet.